Added Error Handling for Internet Problems

+++ photobooth/server/uploader.py
@@ -9,7 +9,7 @@ import os
 import flickrConnect as fCon
 import threading
 from pip._vendor.requests.exceptions import ConnectionError
-from requests.exceptions import ConnectionError
+import xml.etree
 
 class uploaderThread(threading.Thread):
     
@@ -30,12 +30,18 @@ class uploaderThread(threading.Thread):
                 try: 
                     fInst = fCon.flickrConnect()
                     resp = fInst.upload(self.path + picture)
-                    print(resp)
-                    print('Remove:' + picture)
-                    os.remove(self.path + picture)
+                   
+                    if resp is not None:
+                        print('Respond:')
+                        xml.etree.ElementTree.dump(resp)
+                        
+                        print('Remove:' + picture)
+                        os.remove(self.path + picture)
                 
-                except ConnectionError as e:
+                except Exception as e:
+                    print('Upload Failed: ')
                     print(e)
+                    
         
     def run(self):
         while not self._stop_req.isSet():
